+++ plugins/modules/certificate.py
@@ -261,9 +261,6 @@ py4ps_version = utils.py4ps_version_check()
 IS_SUPPORTED_PY4PS_VERSION = py4ps_version['supported_version']
 VERSION_ERROR = py4ps_version['unsupported_version_message']
 
-# Application type
-APPLICATION_TYPE = 'Ansible/3.3.0'
-
 
 class PowerStoreCertificate(object):
     """Certificate operations"""
@@ -296,9 +293,7 @@ class PowerStoreCertificate(object):
         if IS_SUPPORTED_PY4PS_VERSION is False:
             self.module.fail_json(msg=VERSION_ERROR)
 
-        self.conn = utils.get_powerstore_connection(
-            self.module.params,
-            application_type=APPLICATION_TYPE)
+        self.conn = utils.get_powerstore_connection(self.module.params)
         self.configuration = self.conn.config_mgmt
         msg = 'Got Py4ps instance for configuration on' \
               ' PowerStore {0}'.format(self.configuration)