+++ plugins/modules/file_dns.py
@@ -183,9 +183,6 @@ from ansible_collections.dellemc.powerstore.plugins.module_utils.storage.dell.li
 
 LOG = utils.get_logger('file_dns')
 
-# Application type
-APPLICATION_TYPE = 'Ansible/3.3.0'
-
 
 class PowerStoreFileDNS(PowerStoreBase):
     """Class with File DNS Operations"""