+++ plugins/modules/file_interface.py
@@ -216,9 +216,6 @@ from ansible_collections.dellemc.powerstore.plugins.module_utils.storage.dell.li
 
 LOG = utils.get_logger('file_interface')
 
-# Application type
-APPLICATION_TYPE = 'Ansible/3.3.0'
-
 
 class PowerStoreFileInterface(PowerStoreBase):
     """Class with File Interface Operations"""