+++ plugins/modules/info.py
@@ -1909,9 +1909,6 @@ py4ps_version = utils.py4ps_version_check()
 IS_SUPPORTED_PY4PS_VERSION = py4ps_version['supported_version']
 VERSION_ERROR = py4ps_version['unsupported_version_message']
 
-# Application type
-APPLICATION_TYPE = 'Ansible/3.3.0'
-
 
 class PowerstoreInfo(object):
     """Info operations"""
@@ -1945,9 +1942,7 @@ class PowerstoreInfo(object):
         if IS_SUPPORTED_PY4PS_VERSION is False:
             self.module.fail_json(msg=VERSION_ERROR)
 
-        self.conn = utils.get_powerstore_connection(
-            self.module.params,
-            application_type=APPLICATION_TYPE)
+        self.conn = utils.get_powerstore_connection(self.module.params)
         self.provisioning = self.conn.provisioning
         self.protection = self.conn.protection
         self.configuration = self.conn.config_mgmt