+++ plugins/modules/nfs.py
@@ -379,9 +379,6 @@ py4ps_version = utils.py4ps_version_check()
 IS_SUPPORTED_PY4PS_VERSION = py4ps_version['supported_version']
 VERSION_ERROR = py4ps_version['unsupported_version_message']
 
-# Application type
-APPLICATION_TYPE = 'Ansible/3.3.0'
-
 
 class PowerStoreNfsExport(object):
     """Class with NFS export operations"""
@@ -427,8 +424,7 @@ class PowerStoreNfsExport(object):
         if not IS_SUPPORTED_PY4PS_VERSION:
             self.module.fail_json(msg=VERSION_ERROR)
 
-        self.py4ps_conn = utils.get_powerstore_connection(
-            self.module.params, application_type=APPLICATION_TYPE)
+        self.py4ps_conn = utils.get_powerstore_connection(self.module.params)
         self.provisioning = self.py4ps_conn.provisioning
         LOG.info('Got Py4ps instance for PowerStore')