+++ plugins/modules/service_config.py
@@ -114,9 +114,6 @@ from ansible_collections.dellemc.powerstore.plugins.module_utils.storage.dell \
 
 LOG = utils.get_logger('service_config')
 
-# Application type
-APPLICATION_TYPE = 'Ansible/3.3.0'
-
 
 class ServiceConfigs(PowerStoreBase):