+++ plugins/modules/smb_server.py
@@ -194,9 +194,6 @@ from ansible_collections.dellemc.powerstore.plugins.module_utils.storage.dell.li
 
 LOG = utils.get_logger('smb_server')
 
-# Application type
-APPLICATION_TYPE = 'Ansible/3.3.0'
-
 
 class PowerStoreSMBServer(PowerStoreBase):
     """Class with SMB server Operations"""