+++ CHANGELOG.rst
@@ -4,13 +4,23 @@ Dellemc.Powerstore Change Logs
 
 .. contents:: Topics
 
+
+v3.4.0
+======
+
+Minor Changes
+-------------
+
+- Added the support for ACL in the SMB Share module.
+- Info module is enhanced to list acls.
+
 v3.3.0
 ======
 
 Minor Changes
 -------------
 
-- Added the support for PowerStore version 4.0.0.0 (Victory Release)
+- Added the support for PowerStore version 4.0.0.0 (Victory Release).
 
 v3.2.0
 ======